# auth-service/docs/swagger.yaml
definitions:
  models.Admin:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      password:
        type: string
      updated_at:
        type: string
    type: object
  models.UpdateAdminRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
info:
  contact: {}
paths:
  /admins:
    get:
      description: List all admin users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Admin'
            type: array
      summary: List all admin users
    post:
      consumes:
      - application/json
      description: Create a new admin user
      operationId: create-order
      parameters:
      - description: Admin details
        in: body
        name: admin
        required: true
        schema:
          $ref: '#/definitions/models.Admin'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Admin'
      summary: Create a new admin user
  /admins/{id}:
    delete:
      description: Delete an admin user by ID
      parameters:
      - description: Admin ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      summary: Delete an admin user by ID
    get:
      description: Get an admin user by ID
      parameters:
      - description: Admin ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Admin'
      summary: Get an admin user by ID
    put:
      consumes:
      - application/json
      description: Update an admin user by ID
      parameters:
      - description: Admin ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated admin details
        in: body
        name: admin
        required: true
        schema:
          $ref: '#/definitions/models.UpdateAdminRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Admin'
      summary: Update an admin user by ID
  /generate-token:
    post:
      consumes:
      - application/json
      description: Generate a Bearer Token using admin credentials and returns the
        token
      produces:
      - application/json
      responses: {}
      summary: Generate a Bearer Token
swagger: "2.0"